requirements.txt
ultralytics>=8.3.0
opencv-python>=4.8.0
flask>=3.0.0
numpy>=1.24.0
twilio>=9.0.0
mysql-connector-python>=9.0.0
python-dotenv>=1.0.0
